09 Aug 2020

+++ controllers/dasfileoperation.js
@@ -1,10 +1,18 @@
+/*
+Controller Module for File Operation
+    - handles all file operations such as download, edit, delete, upload, open, and browsing of drive
+
+@module fileOperation
+@author Nelson Maligro
+@copyright 2020
+@license GPL
+*/
 module.exports = function(app, arrDB){
   var routeduty = require('./routeduty');
   var fs = require('fs');
   var path = require('path');
   var bodyParser = require('body-parser');
   const cookieParser = require('cookie-parser');
-
   const dbhandle = require('./dbhandle');
   const dochandle = require('./dochandle');
   const monitoring = require('./monitoring');
@@ -13,12 +21,13 @@ module.exports = function(app, arrDB){
   const dateformat = require('dateformat');
   var multer = require('multer');
 
+  //initialize url encoding, cookies, and default drive path
   app.use(cookieParser());
   var urlencodedParser = bodyParser.urlencoded({extended:true});
   var drivetmp = "public/drive/", drive = "D:/Drive/";
   dbhandle.settingDis((setting)=>{drivetmp = setting.publicdrive;});
 
-  //list all document classification and tags
+  //list all branches, classifications and tags
   var docClass = []; var docTag = []; var docBr = [];    var grpUsrs = [];
   dbhandle.generateList(arrDB.class, function (res){ docClass = res; });
   dbhandle.generateList(arrDB.tag, function (res){ docTag = res; });
@@ -32,19 +41,21 @@ module.exports = function(app, arrDB){
       filename: function (req, file, callback) { callback(null, file.originalname);}
     });
     var upload = multer({ storage : storage}).single('fileinput');
-    //post file open with params
+    //
+    //---------------------------------- Express app handling starts here --------------------------------------------------
+    //post handle file open with params
     app.post('/fileopen', urlencodedParser, function(req,res){
       utilsdocms.validToken(req, res,  function (decoded, id){
         postFileopen(req, res, id);
       });
     });
-    //post show file attachments
+    //post handle show file attachments
     app.post('/showfile', urlencodedParser, function(req,res){
       utilsdocms.validToken(req, res,   function (decoded, id){
         showFile(req, res, id);
       });
     });
-    //post handle toggle continue routing or new routing slip
+    //get handle download file from current client view
     app.get('/downloadfile/:file/:view', function(req,res){
       utilsdocms.validToken(req, res,  function (decoded, id){
         downloadfile(req, res, id);
@@ -56,19 +67,19 @@ module.exports = function(app, arrDB){
         deletedoc(req, res, id);
       });
     });
-    //post handle edit file incoming
+    //post handle edit file within client incoming view
     app.post('/editincoming', urlencodedParser, function(req,res){
       utilsdocms.validToken(req, res,  function (decoded, id){
         editincoming(req, res, id);
       });
     });
-    //get open file no params
+    //get handle upload file
     app.get('/fileupload', function(req,res){
       utilsdocms.validToken(req, res,  function (decoded, id){
         getFileupload(req, res, id);
       });
     });
-    //post upload file
+    //post handle upload file
     app.post('/fileupload', urlencodedParser, function(req,res){
       utilsdocms.validToken(req, res,  function (decoded, id){
         postFileUpload(req, res, id);
@@ -80,127 +91,124 @@ module.exports = function(app, arrDB){
         delNotiFile(req, res, id);
       });
     });
-    //handle browse drive
+    //post handle browse drive
     app.post('/browsedrive', urlencodedParser, function(req,res){
       utilsdocms.validToken(req, res,  function (decoded, id){
         console.log('browse drive');
-        // Browse folder "directory"
         let newDrive = req.body.path;
-        //console.log(newDrive);
-        //if (newDrive.toUpperCase().includes("D:/DRIVE/")) newDrive = drive.substring(0,drive.length-1);
-        //if (newDrive.toUpperCase().includes(drive.toUpperCase().substring(0,drive.length-1))) {
+        //Check the client path request if it is within the default drive path
+        //****if (newDrive.toUpperCase().includes("D:/DRIVE/")) newDrive = drive.substring(0,drive.length-1);
+        //****if (newDrive.toUpperCase().includes(drive.toUpperCase().substring(0,drive.length-1))) {
         if ((!newDrive.toUpperCase().includes('D:/DRIVE')) && (!newDrive.toUpperCase().includes(drive.toUpperCase()))) return res.json(JSON.stringify([]));
-          if (newDrive.toUpperCase().includes('D:/DRIVE')) {
-            let drivePre = newDrive.substring(0,8);
-            newDrive = newDrive.replace(drivePre,drive.substring(0,drive.length-1));
-          }
-          //console.log(newDrive);
-          let disFiles = utilsdocms.getFiles(newDrive); let arrFiles = [];
-          disFiles.forEach((item, i) => {
-            if ((!item.includes('~')) && (!item.includes('.idxI')) && (!item.includes('.idxD'))) arrFiles.push(item);
-          });
-          var arrBr = {
-            dirs:utilsdocms.getDirs(newDrive),
-            files:arrFiles
-          }
-          res.json(JSON.stringify(arrBr));
+        if (newDrive.toUpperCase().includes('D:/DRIVE')) { //get the subpath from the default drive
+          let drivePre = newDrive.substring(0,8);
+          newDrive = newDrive.replace(drivePre,drive.substring(0,drive.length-1));
+        }
+        //Get directories and all files except temp and index files
+        let disFiles = utilsdocms.getFiles(newDrive); let arrFiles = [];
+        disFiles.forEach((item, i) => {
+          if ((!item.includes('~')) && (!item.includes('.idxI')) && (!item.includes('.idxD'))) arrFiles.push(item);
+        });
+        var arrBr = {
+          dirs:utilsdocms.getDirs(newDrive),
+          files:arrFiles
+        }
+        res.json(JSON.stringify(arrBr));
         //}
       });
     });
-
-    //////////////////////////////////////FUNCTIONS START HERE///////////////////////////////////////////////
+    //
+    //------------------------------------------FUNCTIONS START HERE----------------------------------------------------
     //handle edit incoming documents
     function editincoming(req,res,id){
-        dbhandle.userFind(id, function (user){
-              let filepath = drivetmp+user.group+'/'+req.body.file;
-              //console.log(filepath);
-              monitoring.getOriginator(req.body.file, function(branch){
-                //console.log(branch.toUpperCase(),user.group.toUpperCase())
-                if (branch.toUpperCase() == user.group.toUpperCase()) {
-                  dbhandle.actlogsCreate(id, Date.now(), 'Edit document during routing', req.body.file, req.ip);
-                  if (!fs.existsSync(drive+'Recoverhere/')) fs.mkdirSync(drive+'Recoverhere/');
-                  if (!fs.existsSync(drive+'Recoverhere/'+user.group.toUpperCase())) fs.mkdirSync(drive+'Recoverhere/'+user.group.toUpperCase());
-                  fs.copyFileSync(path.resolve(filepath),drive+'Recoverhere/'+user.group.toUpperCase()+'/'+req.body.file);
-                  res.json(user.path+'Recoverhere/'+user.group.toUpperCase()+'/'+req.body.file);
-                } else res.json('notowner');
+      dbhandle.userFind(id, function (user){
+        let filepath = drivetmp+user.group+'/'+req.body.file;
+        monitoring.getOriginator(req.body.file, function(branch){
+          if (branch.toUpperCase() == user.group.toUpperCase()) { //if the requesting client is the originator of the doc
+            dbhandle.actlogsCreate(id, Date.now(), 'Edit document during routing', req.body.file, req.ip);
+            if (!fs.existsSync(drive+'Recoverhere/')) fs.mkdirSync(drive+'Recoverhere/');
+            if (!fs.existsSync(drive+'Recoverhere/'+user.group.toUpperCase())) fs.mkdirSync(drive+'Recoverhere/'+user.group.toUpperCase());
+            fs.copyFileSync(path.resolve(filepath),drive+'Recoverhere/'+user.group.toUpperCase()+'/'+req.body.file);
+            res.json(user.path+'Recoverhere/'+user.group.toUpperCase()+'/'+req.body.file);
+          } else res.json('notowner');
 
-              });
-            console.log('edit incoming document');
-          });
+        });
+        console.log('edit incoming document');
+      });
     }
     //handle delete documents
     function deletedoc(req,res,id){
-        dbhandle.userFind(req.body.user, function (user){
-            var filepath = "";
-            if (req.body.branch=="fileopen") filepath = req.body.filepath;
-            else {
-              if (req.body.branch.toUpperCase()=="RELEASE") filepath= drivetmp+'Release/'+req.body.filepath;
-              else filepath= drivetmp+req.body.branch+'/'+req.body.filepath;
-            }
-            //console.log(filepath);
-            if (req.body.branch!="fileopen"){ //if routing
-              monitoring.getOriginator(req.body.filename, function(branch){
-                monitoring.findLastBranch(req.body.filename, user.group, function(found){
-                  //console.log(branch.toUpperCase() + req.body.branch.toUpperCase());
-                  if (!found)  { //All Branches - not part of routing
-                    if (branch.toString().toUpperCase().includes(user.group.toUpperCase())) { //if originator
-                      dbhandle.monitorFindFile(req.body.filename, function(result){ //delete in monitoring
-                        if (result) dbhandle.monitorDel(req.body.filename, function(){});
-                      });
-                    }
-                    dbhandle.docFind(filepath, function(docres){ //delete in pndocs
-                      if (docres) dbhandle.docDel(filepath,()=>{});
-                    });
-                    if (fs.existsSync(filepath)){
-                      if (!fs.existsSync(drive+'Recoverhere/')) fs.mkdirSync(drive+'Recoverhere/');
-                      fs.copyFileSync(filepath,drive+'Recoverhere/'+req.body.filename)
-                      fs.unlink(filepath, (err)=>{if (err) console.log(err);});
-                    }
-                      res.json('successful');
-                      console.log(req.body.branch.toUpperCase());
-                    if (req.body.branch.toUpperCase()=="RELEASE")  { //remove file from monitoring and temp monitoring
-                      dbhandle.actlogsCreate(id, Date.now(), 'Released document deleted', req.body.filename, req.ip);
-                      dbhandle.monitorFindFile(req.body.filename, (result)=>{
-                        if (result) dbhandle.monitorDel(result.filename,()=>{});
-                      });
-                      dbhandle.tempmonitorFindFile(req.body.filename, function(tempresult){
-                        if (tempresult) dbhandle.tempmonitorDel(req.body.filename,()=>{});
-                      });
-                    } else dbhandle.actlogsCreate(id, Date.now(), 'Delete document during routing - branch for info', req.body.filename, req.ip);
-                  } else { //part of routing
-                    if (branch.toString().toUpperCase().includes(user.group.toUpperCase())) { //if originator
-                      dbhandle.monitorFindFile(req.body.filename, function(result){ //delete in monitoring
-                        if (result) dbhandle.monitorDel(req.body.filename, function(){});
-                      });
-                      dbhandle.docFind(filepath, function(docres){ //delete in pndocs
-                        if (docres) dbhandle.docDel(filepath,()=>{});
-                      });
-                      if (fs.existsSync(filepath)){
-                        if (!fs.existsSync(drive+'Recoverhere/')) fs.mkdirSync(drive+'Recoverhere/');
-                        fs.copyFileSync(filepath,drive+'Recoverhere/'+req.body.filename)
-                        fs.unlink(filepath, (err)=>{if (err) console.log(err);});
-                      }
-                      res.json('successful');
-                      dbhandle.actlogsCreate(id, Date.now(), 'Delete document during routing - branch routed', req.body.filename, req.ip);
-                    } else res.json('lastbranch');
-                  }
-
+      dbhandle.userFind(req.body.user, function (user){
+        var filepath = "";
+        if (req.body.branch=="fileopen") filepath = req.body.filepath;
+        else {
+          if (req.body.branch.toUpperCase()=="RELEASE") filepath= drivetmp+'Release/'+req.body.filepath;
+          else filepath= drivetmp+req.body.branch+'/'+req.body.filepath;
+        }
+        //console.log(filepath);
+        if (req.body.branch!="fileopen"){ //if routing
+          monitoring.getOriginator(req.body.filename, function(branch){
+            monitoring.findLastBranch(req.body.filename, user.group, function(found){
+              //console.log(branch.toUpperCase() + req.body.branch.toUpperCase());
+              if (!found)  { //All Branches - not part of routing
+                if (branch.toString().toUpperCase().includes(user.group.toUpperCase())) { //if originator
+                  dbhandle.monitorFindFile(req.body.filename, function(result){ //delete in monitoring
+                    if (result) dbhandle.monitorDel(req.body.filename, function(){});
+                  });
+                }
+                dbhandle.docFind(filepath, function(docres){ //delete in pndocs
+                  if (docres) dbhandle.docDel(filepath,()=>{});
                 });
-              });
-            } else {
-              dbhandle.docFind(filepath, function(docres){ //delete in pndocs
-                if (docres) dbhandle.docDel(filepath,()=>{});
-              });
-              if (fs.existsSync(filepath)){
-                if (!fs.existsSync(drive+'Recoverhere/')) fs.mkdirSync(drive+'Recoverhere/');
-                fs.copyFileSync(filepath,drive+'Recoverhere/'+req.body.filename)
-                fs.unlink(filepath, (err)=>{if (err) console.log(err);});
+                if (fs.existsSync(filepath)){
+                  if (!fs.existsSync(drive+'Recoverhere/')) fs.mkdirSync(drive+'Recoverhere/');
+                  fs.copyFileSync(filepath,drive+'Recoverhere/'+req.body.filename)
+                  fs.unlink(filepath, (err)=>{if (err) console.log(err);});
+                }
+                res.json('successful');
+                console.log(req.body.branch.toUpperCase());
+                if (req.body.branch.toUpperCase()=="RELEASE")  { //remove file from monitoring and temp monitoring
+                  dbhandle.actlogsCreate(id, Date.now(), 'Released document deleted', req.body.filename, req.ip);
+                  dbhandle.monitorFindFile(req.body.filename, (result)=>{
+                    if (result) dbhandle.monitorDel(result.filename,()=>{});
+                  });
+                  dbhandle.tempmonitorFindFile(req.body.filename, function(tempresult){
+                    if (tempresult) dbhandle.tempmonitorDel(req.body.filename,()=>{});
+                  });
+                } else dbhandle.actlogsCreate(id, Date.now(), 'Delete document during routing - branch for info', req.body.filename, req.ip);
+              } else { //part of routing
+                if (branch.toString().toUpperCase().includes(user.group.toUpperCase())) { //if originator
+                  dbhandle.monitorFindFile(req.body.filename, function(result){ //delete in monitoring
+                    if (result) dbhandle.monitorDel(req.body.filename, function(){});
+                  });
+                  dbhandle.docFind(filepath, function(docres){ //delete in pndocs
+                    if (docres) dbhandle.docDel(filepath,()=>{});
+                  });
+                  if (fs.existsSync(filepath)){
+                    if (!fs.existsSync(drive+'Recoverhere/')) fs.mkdirSync(drive+'Recoverhere/');
+                    fs.copyFileSync(filepath,drive+'Recoverhere/'+req.body.filename)
+                    fs.unlink(filepath, (err)=>{if (err) console.log(err);});
+                  }
+                  res.json('successful');
+                  dbhandle.actlogsCreate(id, Date.now(), 'Delete document during routing - branch routed', req.body.filename, req.ip);
+                } else res.json('lastbranch');
               }
-              res.json('successful');
-              dbhandle.actlogsCreate(id, Date.now(), 'Delete document from the File Server', req.body.filename, req.ip);
-            }
-            console.log('deleting document');
+
+            });
           });
+        } else {
+          dbhandle.docFind(filepath, function(docres){ //delete in pndocs
+            if (docres) dbhandle.docDel(filepath,()=>{});
+          });
+          if (fs.existsSync(filepath)){
+            if (!fs.existsSync(drive+'Recoverhere/')) fs.mkdirSync(drive+'Recoverhere/');
+            fs.copyFileSync(filepath,drive+'Recoverhere/'+req.body.filename)
+            fs.unlink(filepath, (err)=>{if (err) console.log(err);});
+          }
+          res.json('successful');
+          dbhandle.actlogsCreate(id, Date.now(), 'Delete document from the File Server', req.body.filename, req.ip);
+        }
+        console.log('deleting document');
+      });
     }
     //handle file download
     function downloadfile(req, res, id){
@@ -238,10 +246,12 @@ module.exports = function(app, arrDB){
             if (req.cookies.fileAI.trim()!=''){
               console.log(path.resolve(drivetmp +'Uploads/' + req.cookies.fileAI.trim()));
               if (fs.existsSync(path.resolve(drivetmp +'Uploads/' + req.cookies.fileAI.trim()))) {
-                fs.copyFileSync(drivetmp +'Uploads/' + req.cookies.fileAI.trim(), req.cookies.realpath + req.cookies.fileAI.trim());
-                fs.unlinkSync(drivetmp +'Uploads/' + req.cookies.fileAI.trim());
-                res.json('successful');
-                dbhandle.actlogsCreate(id, Date.now(), 'Upload File', req.cookies.fileAI.trim(), req.ip);
+                if (req.cookies.realpath.toUpperCase().includes(drive.toUpperCase())){
+                  fs.copyFileSync(drivetmp +'Uploads/' + req.cookies.fileAI.trim(), req.cookies.realpath + req.cookies.fileAI.trim());
+                  fs.unlinkSync(drivetmp +'Uploads/' + req.cookies.fileAI.trim());
+                  res.json('successful');
+                  dbhandle.actlogsCreate(id, Date.now(), 'Upload File', req.cookies.fileAI.trim(), req.ip);
+                } else res.json('error');
               } else res.json('error');
             } else res.json('error');
           })
@@ -320,7 +330,7 @@ module.exports = function(app, arrDB){
                 if (dochandle.getExtension(disFile)!='.pdf'){
                   if (fs.existsSync(drivetmp + 'PDF-temp/'+ disFile +'.pdf')) fs.unlinkSync(drivetmp + 'PDF-temp/'+ disFile +'.pdf');
                   dochandle.convDoctoPDF(disPath+disFile, drivetmp + 'PDF-temp/'+ disFile +'.pdf', function(){
-                   var arrBr = [{disComm:disComm, openpath:user.path, realpath:disPath, path:disDrive + 'PDF-temp/'+ disFile +'.pdf',files:sortArr,disp:disFile,branch:user.group,docClass:docClass, docTag:docTag, rout:rout, ref:ref, enc:enc, disClas:disClas, disTag:disTag}];
+                    var arrBr = [{disComm:disComm, openpath:user.path, realpath:disPath, path:disDrive + 'PDF-temp/'+ disFile +'.pdf',files:sortArr,disp:disFile,branch:user.group,docClass:docClass, docTag:docTag, rout:rout, ref:ref, enc:enc, disClas:disClas, disTag:disTag}];
                     res.json(JSON.stringify(arrBr));
                   });
                 }else {
